[ADD] mission giver, objectives

+++ Assets/Scripts/Resources/ResourceHandle.cs
@@ -4,12 +4,12 @@ public class ResourceHandle : MonoBehaviour
 {
     public ResourceSystem resourceSystem;
 
-    public void ChangeValue(float delta)
+    public void ChangeValue(int delta)
     {
         resourceSystem.ChangeValue(delta);
     }
     
-    public void SetValue(float newValue)
+    public void SetValue(int newValue)
     {
         resourceSystem.SetValue(newValue);
     }
